backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7
httpx==0.25.2
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0

# AI相关（可选）
google-generativeai
face-recognition==1.3.0
numpy==1.26.2

# 邮件服务（可选，如果使用SendGrid）
# sendgrid==6.10.0
